# Change ArrayEach CB Signature to Return Error