--- a/Dockerfile
+++ b/Dockerfile
@@ -1,12 +1,17 @@
-FROM golang:1.6
+FROM golang:1.10
 
-RUN go get github.com/Jeffail/gabs
-RUN go get github.com/bitly/go-simplejson
-RUN go get github.com/pquerna/ffjson
-RUN go get github.com/antonholmquist/jason
-RUN go get github.com/mreiferson/go-ujson
-RUN go get -tags=unsafe -u github.com/ugorji/go/codec
-RUN go get github.com/mailru/easyjson
+ARG PACKAGES='\
+    github.com/a8m/djson \
+    github.com/antonholmquist/jason \
+    github.com/bitly/go-simplejson \
+    github.com/Jeffail/gabs \
+    github.com/mailru/easyjson \
+    github.com/mreiferson/go-ujson \
+    github.com/pquerna/ffjson \
+    github.com/ugorji/go/codec \
+'
+
+RUN go get $PACKAGES
 
 WORKDIR /go/src/github.com/buger/jsonparser
 ADD . /go/src/github.com/buger/jsonparser