--- a/README.md
+++ b/README.md
@@ -56,8 +56,9 @@ if value, err := jsonparser.GetInt(data, "company", "size"); err == nil {
 }
 
 // You can use `ArrayEach` helper to iterate items [item1, item2 .... itemN]
-jsonparser.ArrayEach(data, func(value []byte, dataType jsonparser.ValueType, offset int, err error) {
+jsonparser.ArrayEach(data, func(value []byte, dataType jsonparser.ValueType, offset int) error {
 	fmt.Println(jsonparser.Get(value, "url"))
+  return nil
 }, "person", "avatars")
 
 // Or use can access fields by index!
@@ -152,9 +153,9 @@ If key data type do not match, it will return error.
 
 ### **`ArrayEach`**
 ```go
-func ArrayEach(data []byte, cb func(value []byte, dataType jsonparser.ValueType, offset int, err error), keys ...string)
+func ArrayEach(data []byte, cb func(value []byte, dataType jsonparser.ValueType, offset int) error, keys ...string) (offset int, err error)
 ```
-Needed for iterating arrays, accepts a callback function with the same return arguments as `Get`.
+Needed for iterating arrays, accepts a callback function with the arguments returned by `Get`. Ends iteration immediately if the callback returns an non-nil error.
 
 ### **`ObjectEach`**
 ```go