… benchmark/benchmark_codecgen.go