@vanackere 2020-02-19
@buger : I was in the process of writing the exact same pull request when I noticed this was already there... Any chance to have this merged ?